refactor: moving httpError from types.ts to index.ts
