+++ lib/util/http/types.ts
@@ -1,8 +1,7 @@
 import type { IncomingHttpHeaders } from 'http';
-import {
+import type {
   OptionsOfBufferResponseBody,
   OptionsOfJSONResponseBody,
-  RequestError as RequestError_,
 } from 'got';
 
 export type GotContextOptions = {
@@ -13,8 +12,6 @@ export type GotContextOptions = {
 export type GotOptions = GotBufferOptions | GotJSONOptions;
 export type GotBufferOptions = OptionsOfBufferResponseBody & GotExtraOptions;
 export type GotJSONOptions = OptionsOfJSONResponseBody & GotExtraOptions;
-
-export { RequestError_ as HttpError };
 
 export type GotExtraOptions = {
   abortOnError?: boolean;
